+++ client/api/types/sitemap.ts
@@ -1,0 +1,7 @@
+import { ApiModel } from '@/api'
+
+export interface Response {
+    photos?: ApiModel.SiteMap[]
+    objects?: ApiModel.SiteMap[]
+    events?: ApiModel.SiteMap[]
+}